requirements.txt
tensorflow
numpy
matplotlib
streamlit
pillow
scikit-learn
